[backend] Update backend interface with recovery
