+++ src/moonlink_backend/src/file_utils.rs
@@ -1,0 +1,89 @@
+use crate::error::{Error, Result};
+use moonlink::{ObjectStorageCache, ObjectStorageCacheConfig};
+
+use more_asserts as ma;
+use std::io::ErrorKind;
+
+/// Default local filesystem directory under the above base directory (which defaults to `PGDATA/pg_mooncake`) where all temporary files (used for union read) will be stored under.
+/// The whole directory is cleaned up at moonlink backend start, to prevent file leak.
+pub const DEFAULT_MOONLINK_TEMP_FILE_PATH: &str = "./temp/";
+/// Default object storage read-through cache directory under the above mooncake directory (which defaults to `PGDATA/pg_mooncake`).
+/// The whole directory is cleaned up at moonlink backend start, to prevent file leak.
+pub const DEFAULT_MOONLINK_OBJECT_STORAGE_CACHE_PATH: &str = "./read_through_cache/";
+/// Min left disk space for on-disk cache of the filesystem which cache directory is mounted on.
+const MIN_DISK_SPACE_FOR_CACHE: u64 = 1 << 30; // 1GiB
+
+/// Get temporary directory under base path.
+pub(super) fn get_temp_file_directory_under_base(base_path: &str) -> std::path::PathBuf {
+    std::path::PathBuf::from(base_path).join(DEFAULT_MOONLINK_TEMP_FILE_PATH)
+}
+/// Get cache directory under base path.
+pub(super) fn get_cache_directory_under_base(base_path: &str) -> std::path::PathBuf {
+    std::path::PathBuf::from(base_path).join(DEFAULT_MOONLINK_OBJECT_STORAGE_CACHE_PATH)
+}
+
+/// Util function to get filesystem size for cache directory
+fn get_cache_filesystem_size(path: &str) -> u64 {
+    let vfs_stat = nix::sys::statvfs::statvfs(path).unwrap();
+    let block_size = vfs_stat.block_size();
+    let avai_blocks = vfs_stat.files_available();
+
+    (block_size as u64).checked_mul(avai_blocks as u64).unwrap()
+}
+
+/// Create default object storage cache.
+/// Precondition: cache directory has been created beforehand.
+pub(super) fn create_default_object_storage_cache(
+    cache_directory_pathbuf: std::path::PathBuf,
+) -> ObjectStorageCache {
+    let cache_directory = cache_directory_pathbuf.to_str().unwrap().to_string();
+    let filesystem_size = get_cache_filesystem_size(&cache_directory);
+    ma::assert_ge!(filesystem_size, MIN_DISK_SPACE_FOR_CACHE);
+
+    let cache_config = ObjectStorageCacheConfig {
+        max_bytes: filesystem_size - MIN_DISK_SPACE_FOR_CACHE,
+        cache_directory,
+        optimize_local_filesystem: true,
+    };
+    ObjectStorageCache::new(cache_config)
+}
+
+/// Util function to delete and re-create the given directory.
+pub fn recreate_directory(dir: &str) -> Result<()> {
+    // Clean up directory to place moonlink temporary files.
+    match std::fs::remove_dir_all(dir) {
+        Ok(()) => {}
+        Err(e) => {
+            if e.kind() != ErrorKind::NotFound {
+                return Err(Error::Io(e));
+            }
+        }
+    }
+    std::fs::create_dir_all(dir)?;
+
+    Ok(())
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    use tempfile::TempDir;
+
+    #[test]
+    fn test_recreate_directory() {
+        let tmp = TempDir::new().unwrap();
+        let file = tmp.path().join("tmp.txt");
+        std::fs::write(&file, b"x").unwrap();
+        assert!(file.exists());
+
+        // idempotent "wipe" of an existing dir
+        recreate_directory(tmp.path().to_str().unwrap()).unwrap();
+        assert!(!file.exists());
+
+        // creation of a brand-new path
+        let inner = tmp.path().join("sub");
+        recreate_directory(inner.to_str().unwrap()).unwrap();
+        assert!(inner.exists());
+    }
+}
